refactor: move model handler contracts to agent types
