+++ src/agent/core/flows/CycleServices.ts
@@ -8,11 +8,8 @@ import type { AgentWorkspaceState } from '@agent/core/state/AgentWorkspaceState'
 import type { IToolRegistry } from '@agent/core/tools/ToolTypes';
 import type { IToolUseSession } from '@agent/core/flows/IToolUseSession';
 import type { BaseFlowContextInit } from '@agent/core/flows/BaseFlowServices';
-import type {
-  IModelHandler,
-  SdkToolCall,
-} from '@agent/modelHandlers/types/IModelHandler';
-import type { ProviderMessage } from '@agent/modelHandlers/types/ProviderMessage';
+import type { IModelHandler, SdkToolCall } from '@agent/types/IModelHandler';
+import type { ProviderMessage } from '@agent/types/ProviderMessage';
 import type { TaskRunFileService } from '@utils/files';
 
 /**
