+++ src/agent/core/tools/toolAttachmentExtraction.ts
@@ -1,0 +1,96 @@
+// Local imports - tool result schemas
+import {
+  type FileReference,
+  type ToolFileAttachment,
+  ToolFileAttachmentSchema,
+  type ToolResult,
+  ToolResultSchema,
+  DIAGNOSTIC_TYPE_VALIDATION_ERROR,
+} from '@shared/schemas/toolResult';
+
+const ERROR_PAYLOAD_STRIPPED_KEYS = new Set([
+  'output',
+  'summary',
+  'lineChanges',
+  'edits',
+  'files',
+  'editedFiles',
+]);
+
+/**
+ * Result from extracting attachments from a tool result.
+ * Simple interface - no runtime validation needed for this structure.
+ */
+export interface ExtractedToolAttachments {
+  /** Extracted file attachments with binary data. */
+  attachments: ToolFileAttachment[];
+  /** Sanitized result payload without binary data. */
+  sanitizedResult: ToolResult;
+}
+
+/**
+ * Type guard to check if a value is a valid ToolFileAttachment.
+ * Uses Zod schema for validation.
+ */
+function isToolFileAttachment(value: unknown): value is ToolFileAttachment {
+  return ToolFileAttachmentSchema.safeParse(value).success;
+}
+
+/**
+ * Extracts file attachments from a tool result and returns a typed payload.
+ * Binary data (base64Data, bytes) is stripped from the result.
+ *
+ * Uses the source-level `ToolResultSchema` discriminator; tools declare
+ * success vs error before this projection sees the result.
+ *
+ * @param result - Raw tool result, possibly containing binary data.
+ * @returns Extracted attachments and typed payload without binary data.
+ */
+export function extractToolAttachments(
+  result: ToolResult,
+): ExtractedToolAttachments {
+  const attachmentsCandidate = result.files;
+  const attachments: ToolFileAttachment[] = Array.isArray(attachmentsCandidate)
+    ? attachmentsCandidate.filter(isToolFileAttachment)
+    : [];
+
+  const parsed = ToolResultSchema.parse(result);
+  const status = parsed.status;
+  const sanitizedResult: Record<string, unknown> = { status };
+
+  for (const [key, value] of Object.entries(parsed)) {
+    if (value === undefined) continue;
+    if (key === 'base64Data' || key === 'bytes') {
+      continue;
+    }
+    if (status === 'executed' && key === 'error') {
+      continue;
+    }
+    if (status === 'error' && ERROR_PAYLOAD_STRIPPED_KEYS.has(key)) {
+      continue;
+    }
+    if (key === 'diagnostics' && value && typeof value === 'object') {
+      const diag = value as Record<string, unknown>;
+      if (diag.type === DIAGNOSTIC_TYPE_VALIDATION_ERROR && diag.formatted) {
+        sanitizedResult[key] = { type: diag.type, formatted: diag.formatted };
+      }
+      continue;
+    }
+    sanitizedResult[key] = value;
+  }
+
+  if (status === 'executed' && attachments.length > 0) {
+    sanitizedResult.files = attachments.map((file): FileReference => ({
+      path: file.path,
+      mimeType: file.mimeType,
+      ...(file.description ? { description: file.description } : {}),
+    }));
+  } else {
+    delete sanitizedResult.files;
+  }
+
+  return {
+    attachments,
+    sanitizedResult: ToolResultSchema.parse(sanitizedResult),
+  };
+}
