+++ src/agent/implementations/flows/reflection/nodes/PrepareContextNode.ts
@@ -1,7 +1,7 @@
 import { Node } from '@agent/node';
 import { ConversationRoundStateSnapshotSchema } from '@agent/core/state/AgentState';
 import { FlowTransition } from '@agent/core/flows/FlowTransitions';
-import type { ProviderMessage } from '@agent/modelHandlers/types/ProviderMessage';
+import type { ProviderMessage } from '@agent/types/ProviderMessage';
 import { appendCompileFailureRoundContext } from '@agent/output/compileFailureRoundContext';
 
 import type { FlowParams } from '@agent/core/flows/BaseFlowServices';
