+++ src/agent/types/ModelHandlerContracts.ts
