+++ src/test-kernel/agent/ToolUseDispatchParallel.vitest.ts
@@ -13,7 +13,7 @@ import { AgentRunStateSnapshotSchema } from '@agent/core/state/AgentState';
 import { AgentWorkspaceState } from '@agent/core/state/AgentWorkspaceState';
 import { MapToolRegistry } from '@agent/core/tools/ToolTypes';
 import type { ITool } from '@agent/core/tools/ToolTypes';
-import type { SdkToolCall } from '@agent/modelHandlers/types/IModelHandler';
+import type { SdkToolCall } from '@agent/types/IModelHandler';
 import { noopAgentRuntimeHost } from '@agent/runtime/AgentRuntimeHost';
 import type { StreamTabId } from '@shared/schemas';
 import type { ToolResult } from '@shared/schemas/toolResult';
